refactor(ffmpeg): move runtime path-correctness into the Detector (#2 Phase B-2)
Completes the capability extraction: the exec-coupled path-correctness cluster
(PreflightPathCorrectness, test*/measure* path probes, observeRuntimePathCorrectness,
updateRuntimePathCapability) + its state (pathProbeFn, pathCorrectnessChecked,
signalStatsYAvgFn) move into the Detector. The two adapter process hooks it needs
(recordProcessDetail, terminateProcessGroup) become Detector callback fields, set
post-construction; HLSRoot becomes a Detector dep.

- adapter.go 1679 -> 1459 (2273 -> 1459 across all of #2, -36%)
- detector.go now owns all encoder + path capability detection; adapter keeps a
  thin PreflightPathCorrectness delegator for bootstrap and calls
  a.detector.observeRuntimePathCorrectness from the exec path

Pure move, no behavior change: full module build + go vet clean, ffmpeg argv
golden tests green, bootstrap wiring test green, golangci-lint --new-from-rev = 0.
